fixup the debug break macro
allow SDL to be excluded

fixup inline and missing typename for gcc

turnoff SDL by default for makefile builds

turn off pch by default; try to use multiple jobs

use snprintf

add a dummy impl

Fix compilation under gcc and clang.

+++ test/bklib/dictionary_test.cpp
@@ -9,7 +9,7 @@
 namespace {
 
 struct test_def {
-    using id_type = bklib::tagged_value<uint32_t, test_def>;
+    using id_type = bklib::tagged_value<size_t, test_def>;
 
     test_def(bklib::utf8_string id_string, int const data)
       : id {bklib::hash_value(id_string)}
